\name{cvarpyp-package}
\alias{cvarpyp-package}
\alias{cvarpyp}
\docType{package}
\title{
\packageTitle{cvarpyp}
}
\description{
\packageDescription{cvarpyp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cvarpyp}
\packageIndices{cvarpyp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cvarpyp}

Maintainer: \packageMaintainer{cvarpyp}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
