% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{get_post_summary_vc}
\alias{get_post_summary_vc}
\title{Get the posterior results of varying coefficients}
\usage{
get_post_summary_vc(
  time_range,
  knots,
  gamma.sample,
  theta.sample,
  grids = 50,
  burn = NULL
)
}
\description{
Get the posterior results of varying coefficients
}
