% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{summarise_beta}
\alias{summarise_beta}
\title{Draw the posterior results of fixed effects}
\usage{
summarise_beta(beta.sample, position, burn = NULL)
}
\description{
Draw the posterior results of fixed effects
}
