% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{summarise_cluster}
\alias{summarise_cluster}
\title{Draw the posterior results of clusters}
\usage{
summarise_cluster(cluster.sample, burn = NULL)
}
\description{
Draw the posterior results of clusters
}
