% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{summarise_latent_location}
\alias{summarise_latent_location}
\title{Draw the posterior results of latent location parameters}
\usage{
summarise_latent_location(
  time_range,
  knots,
  gamma.sample,
  cluster_number,
  variable_number,
  burn = NULL
)
}
\description{
Draw the posterior results of latent location parameters
}
