% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{summarise_Psi}
\alias{summarise_Psi}
\title{Draw the posterior results of random effects}
\usage{
summarise_Psi(Psi.sample, row.position, col.position, burn = NULL)
}
\description{
Draw the posterior results of random effects
}
