% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{summarise_varying_coefficient}
\alias{summarise_varying_coefficient}
\title{Draw the posterior results of varying coefficients}
\usage{
summarise_varying_coefficient(vc_object, cluster_number, variable_number)
}
\description{
Draw the posterior results of varying coefficients
}
